requirements.txt
fastapi
uvicorn
openai
chromadb
python-dotenv
pandas
pydantic
httpx
tenacity
